move stacked to case components
