+++ apps/consultation-portal/screens/Case/components/CaseDocuments/CaseDocuments.tsx
@@ -1,8 +1,8 @@
 import { CardSkeleton } from '../../../../components'
 import { Document } from '../../../../types/interfaces'
-import StackedTitleAndDescription from '../../../../components/StackedTitleAndDescription/StackedTitleAndDescription'
 import { LinkV2 } from '@island.is/island-ui/core'
 import env from '../../../../lib/environment'
+import { Stacked } from '../../components'
 
 interface Props {
   title: string
@@ -12,7 +12,7 @@ interface Props {
 export const CaseDocuments = ({ title, documents }: Props) => {
   return (
     <CardSkeleton>
-      <StackedTitleAndDescription title={title}>
+      <Stacked title={title}>
         {documents.map((document, index) => {
           return (
             <LinkV2
@@ -26,7 +26,7 @@ export const CaseDocuments = ({ title, documents }: Props) => {
             </LinkV2>
           )
         })}
-      </StackedTitleAndDescription>
+      </Stacked>
     </CardSkeleton>
   )
 }
