+++ apps/consultation-portal/screens/Case/components/CaseStatusBox/CaseStatusBox.tsx
@@ -1,5 +1,5 @@
 import { CardSkeleton } from '../../../../components'
-import StackedTitleAndDescription from '../../../../components/StackedTitleAndDescription/StackedTitleAndDescription'
+import StackedTitleAndDescription from '../Stacked/Stacked'
 import { Box, Button, Text } from '@island.is/island-ui/core'
 import Link from 'next/link'
 import localization from '../../Case.json'
