+++ apps/consultation-portal/screens/Case/components/Coordinator/Coordinator.tsx
@@ -1,5 +1,5 @@
 import { CardSkeleton } from '../../../../components'
-import StackedTitleAndDescription from '../../../../components/StackedTitleAndDescription/StackedTitleAndDescription'
+import StackedTitleAndDescription from '../Stacked/Stacked'
 import { Text } from '@island.is/island-ui/core'
 import localization from '../../Case.json'
