+++ apps/consultation-portal/screens/Case/components/index.ts
@@ -1,4 +1,5 @@
 export { default as CaseEmailBox } from './CaseEmailBox/CaseEmailBox'
+export { default as CaseEmailActionBox } from './CaseEmailBox/CaseEmailActionBox'
 export { default as CaseOverview } from './CaseOverview/CaseOverview'
 export { default as CaseTimeline } from './CaseTimeline/CaseTimeline'
 export { default as Coordinator } from './Coordinator/Coordinator'
@@ -9,3 +10,4 @@ export { default as CaseDocuments } from './CaseDocuments/CaseDocuments'
 export { default as AdviceList } from './AdviceList/AdviceList'
 export { default as AdviceCard } from './AdviceCard/AdviceCard'
 export { default as AdviceSkeletonLoader } from './AdviceSkeletonLoader/AdviceSkeletonLoader'
+export { default as Stacked } from './Stacked/Stacked'
